requirements.txt
Flask==3.1.2
rembg==2.0.67
Pillow==11.3.0
opencv-python==4.12.0.88
numpy==2.3.3
Werkzeug==3.1.3
onnxruntime==1.23.0
